Replaced data models with functions in backend
- used a more imperative approach over OOP for
simplicity
